02-backend/requirements.txt
fastapi
uvicorn[standard]
google-cloud-aiplatform>=1.38.0
google-cloud-firestore
google-cloud-storage
pydantic
pydantic-settings
python-dotenv